requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
psycopg2>=2.9
drf-spectacular>=0.27.1
django-simple-history>=3.3.0
gunicorn>=21.0.0